DRYed a few things up
